Anchor fix
Fixed an issue with the anchor where the player could change the planet's speed to get an anchored planet to move. Now an anchored planet cannot move. The planet's speed and direction can still be changed while the planet is anchored but these changes will not show until the anchor timer runs out. This is intentional.