Added debris, debris spawner and gravity fields
I have created debris prefabs as well as a script to control these objects. I have also created a script that will act as a debris spawner. It creates a pool of x debris and spawns items from that pool every y seconds. This spawner also deals with removing debris as it enters the void and leaves the screen. It will manage the pool to make sure we never spawn more than x debris. The debris spawns on the screen edges with a random speed, position and direction. A curve is created for the debris to follow, this curve avoids the center of the screen. I added a gravity "Zone of influence" onto the planet and the void. Allowing them to attract the debris that flys around the screen.