Debris consumption, expulsion, simple scoring
I implemented a way for a planet to suck up debris. The planet has a max amount of debris is can hold. This is indicated in the UI. The player can expel the debris inwards or outwards (towards the black hold or the screens edge). There is also now a simple scoring system, different debris have different point values.